@@ -21,21 +21,29 @@
   "author": "charles peters <charlespeters42@gmail.com> (https://charlespeters.net)",
   "license": "MIT",
   "dependencies": {
-    "chalk": "^2.4.1",
-    "fs-extra": "^7.0.0",
-    "mri": "^1.1.1",
+    "chalk": "^2.4.2",
+    "fs-extra": "^7.0.1",
+    "mri": "^1.1.4",
     "node-emoji": "^1.5.1"
   },
   "devDependencies": {
+    "@types/del": "^3.0.1",
+    "@types/execa": "^0.9.0",
     "@types/fs-extra": "^5.0.4",
+    "@types/jest": "^23.3.12",
     "@types/mri": "^1.1.0",
-    "@types/node": "^10.12.0",
+    "@types/node": "^10.12.18",
     "@types/node-emoji": "^1.8.0",
     "del": "^3.0.0",
     "execa": "^1.0.0",
     "jest": "^23.6.0",
-    "prettier": "^1.14.3",
-    "typescript": "^3.1.3"
+    "prettier": "^1.15.3",
+    "ts-jest": "^23.10.5",
+    "typescript": "^3.2.2"
+  },
+  "jest": {
+    "preset": "ts-jest",
+    "testEnvironment": "node"
   },
   "prettier": {
     "printWidth": 95,
